Squashed commit of the following:
commit 19b2f32e5e70efa824987087f74ad32c0ded022e
Author: AJ Slater <aj@slater.net>
Date:   Tue Feb 24 18:30:03 2026 -0800

    simplify copy new files

commit b340838e0c1c848102d5b350926ba6523e341dc0
Author: AJ Slater <aj@slater.net>
Date:   Tue Feb 24 16:50:44 2026 -0800

    add scripts

commit e535bc89f1c46588b01334c7ea09fc030d307f23
Author: AJ Slater <aj@slater.net>
Date:   Tue Feb 24 16:40:45 2026 -0800

    order init makefile properly with ci and docker makefiles

commit 25e6051f08422e501d7c3feeb5982640d1bb3bdc
Author: AJ Slater <aj@slater.net>
Date:   Tue Feb 24 16:40:29 2026 -0800

    update devenv

commit 5c67446c8970bfa8c44036aa11cae3a85fecc4a6
Author: AJ Slater <aj@slater.net>
Date:   Tue Feb 24 16:38:22 2026 -0800

    move fix & lint docker scripts and ci scripts to own scripts and makefiles

commit 572ca99cf29d536b27c05882bfd3c2fb5ac9d658
Author: AJ Slater <aj@slater.net>
Date:   Tue Feb 24 16:29:44 2026 -0800

    update deps

commit 0ad8d656dc5ca3f89284db7e337a97911a0a15b1
Author: AJ Slater <aj@slater.net>
Date:   Tue Feb 24 16:29:40 2026 -0800

    remove unused config file

commit d74d0aa34064c65a1d8541d30ed7e81e06ebfe1c
Author: AJ Slater <aj@slater.net>
Date:   Tue Feb 24 16:29:25 2026 -0800

    always call parse-config from update-devenv. only fix and report on files if any were merged.

commit 0d17a0de1803ff47624f6502b07d70e88cafdd24
Author: AJ Slater <aj@slater.net>
Date:   Tue Feb 24 16:28:27 2026 -0800

    only report on copy & merging if anything happened. merge dotfiles must only copy rc and ignore files.

commit 7941d6c33c657fe8f9cc7b87a01a7bdcaa5aa434
Author: AJ Slater <aj@slater.net>
Date:   Tue Feb 24 16:27:26 2026 -0800

    only report on deleted files we did so

commit 5a8673c63b03826a7a751d3f3f81972d5e1079de
Author: AJ Slater <aj@slater.net>
Date:   Tue Feb 24 16:27:01 2026 -0800

    unset parsed DEVENV variables before setting

commit 43ba9227dcf8a4fc447633518f8bb161e5a3f2f2
Author: AJ Slater <aj@slater.net>
Date:   Tue Feb 24 16:26:17 2026 -0800

    fix calling update-devenv with DEVENV_SRC from makefile

commit d4d9f53912bdba49347c2f10909a13bc070929cb
Author: AJ Slater <aj@slater.net>
Date:   Tue Feb 24 16:25:56 2026 -0800

    format

commit 121116af9be7837ebc51323bccec5404356a4090
Author: AJ Slater <aj@slater.net>
Date:   Tue Feb 24 15:04:40 2026 -0800

    put templates and init in subdirs as well. make merge-dotfiles and copy-new-files feature aware

commit a8bb6292fc51ecc691acacfdbcdb1b763f3094de
Author: AJ Slater <aj@slater.net>
Date:   Tue Feb 24 14:32:26 2026 -0800

    only include scripts from declared features. accept new source from config as well

commit c3e723f384bd957ba9891cf23e15ff818ca2476a
Author: AJ Slater <aj@slater.net>
Date:   Tue Feb 24 13:26:25 2026 -0800

    add merge and copy scripts to delete

commit 17d5abc126a9f6f3da541d48d8d57f78deb21727
Author: AJ Slater <aj@slater.net>
Date:   Tue Feb 24 13:26:14 2026 -0800

    use forloop for subdir copy

commit 8914f71a1b85d54f3b983ddc332435036c1aab55
Author: AJ Slater <aj@slater.net>
Date:   Tue Feb 24 13:22:59 2026 -0800

    don't refer to merge scripts locally

commit 2bb8f2c38eeb9c4670111a739486a97c9ab81c83
Author: AJ Slater <aj@slater.net>
Date:   Mon Feb 23 17:24:01 2026 -0800

    a yaml config file. parse features.

commit 51d2c882807fc45311001c9c28e915caada0d7a7
Author: AJ Slater <aj@slater.net>
Date:   Mon Feb 23 17:13:07 2026 -0800

    other makefiles in makefile comments

+++ scripts/delete-files.sh
@@ -9,5 +9,7 @@ while IFS= read -r file || [[ -n "$file" ]]; do
   [[ -f "$file" ]] && existing_files+=("$file")
 done < "$DELETE_FILE"
 
-echo "Deleting ${#existing_files[@]} files..."
-rm -f -- "${existing_files[@]}"
+if ((${#existing_files[@]})); then
+  echo "Deleting ${#existing_files[@]} files..."
+  rm -f -- "${existing_files[@]}"
+fi
